[新增]
1. 會員中心 - 上方餘額顯示部份增加 活期理財錢包, 目前有24小時候沒有結息的問題
2. 活期理財錢包頁面 - 剩餘累計收益金額( 待產品網關提供接口), 轉出至賬戶餘額中的利息時長(平台網關確認後回覆), 點擊banner跳轉活動頁(待凱森提供)
3. 轉賬記錄頁面 - 剩餘利息賬單紀錄, 待產品網關提供接口, 轉出後賬單紀錄中的狀態未更改(平台網關確認後回覆)

+++ Hybird_1e3c3b/Mine/Model/BTTCustomerBalanceModel.h
@@ -29,8 +29,8 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) NSInteger tlbinCredit;
 @property (nonatomic, strong) NSDictionary *walletBalance;
 @property (nonatomic, assign) double withdrawBal;
-@property (nonatomic, assign) NSInteger yebAmount;
-@property (nonatomic, assign) NSInteger yebInterest;
+@property (nonatomic, assign) CGFloat yebAmount;
+@property (nonatomic, assign) CGFloat yebInterest;
 
 @end
